requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
build==1.2.2
    # via pip-tools
certifi==2024.8.30
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via pretty-errors
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distlib==0.3.8
    # via virtualenv
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   virtualenv
fonttools==4.53.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.9.0
    # via
    #   huggingface-hub
    #   lightning
    #   pytorch-lightning
    #   torch
ftfy==6.2.3
    # via open-clip-torch
huggingface-hub==0.25.0
    # via
    #   open-clip-torch
    #   timm
    #   tokenizers
    #   transformers
identify==2.6.1
    # via pre-commit
idna==3.10
    # via
    #   requests
    #   yarl
imgdl==2.0.1
    # via -r requirements.in
jinja2==3.1.4
    # via torch
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.7
    # via matplotlib
lightning==2.4.0
    # via -r requirements.in
lightning-utilities==0.11.7
    # via
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.0
    # via
    #   -r requirements.in
    #   pycocotools
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.3
    # via torch
nodeenv==1.9.1
    # via pre-commit
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pycocotools
    #   scikit-learn
    #   scipy
    #   torchmetrics
    #   torchvision
    #   transformers
open-clip-torch==2.26.1
    # via -r requirements.in
packaging==24.1
    # via
    #   build
    #   huggingface-hub
    #   lightning
    #   lightning-utilities
    #   matplotlib
    #   pytorch-lightning
    #   torchmetrics
    #   transformers
pillow==9.5.0
    # via
    #   imgdl
    #   matplotlib
    #   torchvision
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via virtualenv
pre-commit==3.8.0
    # via -r requirements.in
pretty-errors==1.2.25
    # via torchmetrics
pycocotools==2.0.8
    # via -r requirements.in
pydantic==1.10.18
    # via imgdl
pyparsing==3.1.4
    # via matplotlib
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via matplotlib
python-json-logger==2.0.7
    # via imgdl
pytorch-lightning==2.4.0
    # via lightning
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   lightning
    #   pre-commit
    #   pytorch-lightning
    #   timm
    #   transformers
regex==2024.9.11
    # via
    #   open-clip-torch
    #   transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   imgdl
    #   transformers
safetensors==0.4.5
    # via
    #   timm
    #   transformers
scikit-learn==1.5.0
    # via -r requirements.in
scipy==1.13.1
    # via
    #   -r requirements.in
    #   scikit-learn
six==1.16.0
    # via python-dateutil
sympy==1.13.2
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
timm==1.0.9
    # via open-clip-torch
tokenizers==0.19.1
    # via transformers
tomli==2.0.1
    # via
    #   build
    #   pip-tools
torch==2.4.0
    # via
    #   -r requirements.in
    #   lightning
    #   open-clip-torch
    #   pytorch-lightning
    #   timm
    #   torchmetrics
    #   torchvision
torchmetrics==1.4.0
    # via
    #   -r requirements.in
    #   lightning
    #   pytorch-lightning
torchvision==0.19.0
    # via
    #   -r requirements.in
    #   open-clip-torch
    #   timm
tqdm==4.66.4
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   imgdl
    #   lightning
    #   open-clip-torch
    #   pytorch-lightning
    #   transformers
transformers==4.44.2
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   lightning
    #   lightning-utilities
    #   multidict
    #   pydantic
    #   pytorch-lightning
    #   torch
urllib3==2.2.3
    # via requests
virtualenv==20.26.5
    # via pre-commit
wcwidth==0.2.13
    # via ftfy
wget==3.2
    # via -r requirements.in
wheel==0.44.0
    # via pip-tools
yarl==1.11.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
